// typescript/packages/objects/generated/proto/objects/v1/objects_pb.d.ts
export declare type BucketRef = Message<"acyclic.objects.v1.BucketRef"> & {
  /**
   * @generated from field: string bucket_id = 1;
   */
  bucketId: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;
};
// ...
export declare type ObjectMetadata = Message<"acyclic.objects.v1.ObjectMetadata"> & {
  /**
   * @generated from field: string content_type = 1;
   */
  contentType: string;

  /**
   * @generated from field: map<string, string> user = 2;
   */
  user: { [key: string]: string };

  /**
   * @generated from field: string content_encoding = 3;
   */
  contentEncoding: string;

  /**
   * @generated from field: string cache_control = 4;
   */
  cacheControl: string;

  /**
   * @generated from field: string content_disposition = 5;
   */
  contentDisposition: string;

  /**
   * @generated from field: string content_language = 6;
   */
  contentLanguage: string;

  /**
   * @generated from field: optional int64 expires_unix_seconds = 7;
   */
  expiresUnixSeconds?: bigint | undefined;
};

/**
 * Describes the message acyclic.objects.v1.ObjectMetadata.
 * Use `create(ObjectMetadataSchema)` to create a new message.
 */
export declare const ObjectMetadataSchema: GenMessage<ObjectMetadata>;

/**
 * @generated from message acyclic.objects.v1.Preconditions
 */
export declare type Preconditions = Message<"acyclic.objects.v1.Preconditions"> & {
  /**
   * @generated from oneof acyclic.objects.v1.Preconditions.condition
   */
  condition: {
    /**
     * @generated from field: bool if_absent = 1;
     */
    value: boolean;
    case: "ifAbsent";
  } | {
    /**
     * @generated from field: string if_match = 2;
     */
    value: string;
    case: "ifMatch";
  } | {
    /**
     * @generated from field: string if_version = 3;
     */
    value: string;
    case: "ifVersion";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message acyclic.objects.v1.Preconditions.
 * Use `create(PreconditionsSchema)` to create a new message.
 */
export declare const PreconditionsSchema: GenMessage<Preconditions>;

/**
 * @generated from message acyclic.objects.v1.MutationIdentity
 */
export declare type MutationIdentity = Message<"acyclic.objects.v1.MutationIdentity"> & {
  /**
   * @generated from field: string idempotency_key = 1;
   */
  idempotencyKey: string;
};

/**
 * Describes the message acyclic.objects.v1.MutationIdentity.
 * Use `create(MutationIdentitySchema)` to create a new message.
 */
export declare const MutationIdentitySchema: GenMessage<MutationIdentity>;

/**
 * @generated from message acyclic.objects.v1.Bucket
 */
export declare type Bucket = Message<"acyclic.objects.v1.Bucket"> & {
  /**
   * @generated from field: acyclic.objects.v1.BucketRef bucket = 1;
   */
  bucket?: BucketRef | undefined;

  /**
   * @generated from field: google.protobuf.Timestamp created_at = 2;
   */
  createdAt?: Timestamp | undefined;
};

/**
 * Describes the message acyclic.objects.v1.Bucket.
 * Use `create(BucketSchema)` to create a new message.
 */
export declare const BucketSchema: GenMessage<Bucket>;

/**
 * @generated from message acyclic.objects.v1.CreateBucketRequest
 */
export declare type CreateBucketRequest = Message<"acyclic.objects.v1.CreateBucketRequest"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * @generated from field: acyclic.objects.v1.MutationIdentity mutation = 2;
   */
  mutation?: MutationIdentity | undefined;
};

/**
 * Describes the message acyclic.objects.v1.CreateBucketRequest.
 * Use `create(CreateBucketRequestSchema)` to create a new message.
 */
export declare const CreateBucketRequestSchema: GenMessage<CreateBucketRequest>;

/**
 * @generated from message acyclic.objects.v1.HeadBucketRequest
 */
export declare type HeadBucketRequest = Message<"acyclic.objects.v1.HeadBucketRequest"> & {
  /**
   * @generated from field: acyclic.objects.v1.BucketRef bucket = 1;
   */
  bucket?: BucketRef | undefined;
};

/**
 * Describes the message acyclic.objects.v1.HeadBucketRequest.
 * Use `create(HeadBucketRequestSchema)` to create a new message.
 */
export declare const HeadBucketRequestSchema: GenMessage<HeadBucketRequest>;

/**
 * @generated from message acyclic.objects.v1.DeleteBucketRequest
 */
export declare type DeleteBucketRequest = Message<"acyclic.objects.v1.DeleteBucketRequest"> & {
  /**
   * @generated from field: acyclic.objects.v1.BucketRef bucket = 1;
   */
  bucket?: BucketRef | undefined;

  /**
   * @generated from field: acyclic.objects.v1.MutationIdentity mutation = 2;
   */
  mutation?: MutationIdentity | undefined;
};

/**
 * Describes the message acyclic.objects.v1.DeleteBucketRequest.
 * Use `create(DeleteBucketRequestSchema)` to create a new message.
 */
export declare const DeleteBucketRequestSchema: GenMessage<DeleteBucketRequest>;

/**
 * @generated from message acyclic.objects.v1.DeleteBucketResponse
 */
export declare type DeleteBucketResponse = Message<"acyclic.objects.v1.DeleteBucketResponse"> & {
  /**
   * @generated from field: bool existed = 1;
   */
  existed: boolean;
};

/**
 * Describes the message acyclic.objects.v1.DeleteBucketResponse.
 * Use `create(DeleteBucketResponseSchema)` to create a new message.
 */
export declare const DeleteBucketResponseSchema: GenMessage<DeleteBucketResponse>;

/**
 * @generated from message acyclic.objects.v1.PutObjectHeader
 */
export declare type PutObjectHeader = Message<"acyclic.objects.v1.PutObjectHeader"> & {
  /**
   * @generated from field: acyclic.objects.v1.BucketRef bucket = 1;
   */
  bucket?: BucketRef | undefined;

  /**
   * @generated from field: string object_key = 2;
   */
  objectKey: string;

  /**
   * @generated from field: acyclic.objects.v1.ObjectMetadata metadata = 3;
   */
  metadata?: ObjectMetadata | undefined;

  /**
   * @generated from field: acyclic.objects.v1.Preconditions preconditions = 4;
   */
  preconditions?: Preconditions | undefined;

  /**
   * @generated from field: acyclic.objects.v1.MutationIdentity mutation = 5;
   */
  mutation?: MutationIdentity | undefined;
};

/**
 * Describes the message acyclic.objects.v1.PutObjectHeader.
 * Use `create(PutObjectHeaderSchema)` to create a new message.
 */
export declare const PutObjectHeaderSchema: GenMessage<PutObjectHeader>;

/**
 * @generated from message acyclic.objects.v1.PutObjectRequest
 */
export declare type PutObjectRequest = Message<"acyclic.objects.v1.PutObjectRequest"> & {
  /**
   * @generated from oneof acyclic.objects.v1.PutObjectRequest.frame
   */
  frame: {
    /**
     * @generated from field: acyclic.objects.v1.PutObjectHeader header = 1;
     */
    value: PutObjectHeader;
    case: "header";
  } | {
    /**
     * @generated from field: bytes body = 2;
     */
    value: Uint8Array;
    case: "body";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message acyclic.objects.v1.PutObjectRequest.
 * Use `create(PutObjectRequestSchema)` to create a new message.
 */
export declare const PutObjectRequestSchema: GenMessage<PutObjectRequest>;
// ...
export declare const ObjectVersionSchema: GenMessage<ObjectVersion>;
// ...
export declare const GetObjectRequestSchema: GenMessage<GetObjectRequest>;
// ...
export declare const GetObjectResponseSchema: GenMessage<GetObjectResponse>;
// ...
export declare const HeadObjectRequestSchema: GenMessage<HeadObjectRequest>;
// ...
export declare const HeadObjectResponseSchema: GenMessage<HeadObjectResponse>;
// ...
export declare const DeleteObjectRequestSchema: GenMessage<DeleteObjectRequest>;
// ...
export declare const DeleteObjectResponseSchema: GenMessage<DeleteObjectResponse>;
// ...
export declare const ListObjectsRequestSchema: GenMessage<ListObjectsRequest>;
// ...
export declare const ListObjectsResponseSchema: GenMessage<ListObjectsResponse>;
// ...
export declare const CreateMultipartRequestSchema: GenMessage<CreateMultipartRequest>;

/**
 * @generated from message acyclic.objects.v1.MultipartUpload
 */
export declare type MultipartUpload = Message<"acyclic.objects.v1.MultipartUpload"> & {
  /**
   * @generated from field: string upload_id = 1;
   */
  uploadId: string;
};

/**
 * Describes the message acyclic.objects.v1.MultipartUpload.
 * Use `create(MultipartUploadSchema)` to create a new message.
 */
export declare const MultipartUploadSchema: GenMessage<MultipartUpload>;

/**
 * @generated from message acyclic.objects.v1.UploadPartHeader
 */
export declare type UploadPartHeader = Message<"acyclic.objects.v1.UploadPartHeader"> & {
  /**
   * @generated from field: acyclic.objects.v1.BucketRef bucket = 1;
   */
  bucket?: BucketRef | undefined;

  /**
   * @generated from field: string object_key = 2;
   */
  objectKey: string;

  /**
   * @generated from field: string upload_id = 3;
   */
  uploadId: string;

  /**
   * @generated from field: uint32 part_number = 4;
   */
  partNumber: number;

  /**
   * @generated from field: acyclic.objects.v1.MutationIdentity mutation = 5;
   */
  mutation?: MutationIdentity | undefined;
};

/**
 * Describes the message acyclic.objects.v1.UploadPartHeader.
 * Use `create(UploadPartHeaderSchema)` to create a new message.
 */
export declare const UploadPartHeaderSchema: GenMessage<UploadPartHeader>;

/**
 * @generated from message acyclic.objects.v1.UploadPartRequest
 */
export declare type UploadPartRequest = Message<"acyclic.objects.v1.UploadPartRequest"> & {
  /**
   * @generated from oneof acyclic.objects.v1.UploadPartRequest.frame
   */
  frame: {
    /**
     * @generated from field: acyclic.objects.v1.UploadPartHeader header = 1;
     */
    value: UploadPartHeader;
    case: "header";
  } | {
    /**
     * @generated from field: bytes body = 2;
     */
    value: Uint8Array;
    case: "body";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message acyclic.objects.v1.UploadPartRequest.
 * Use `create(UploadPartRequestSchema)` to create a new message.
 */
export declare const UploadPartRequestSchema: GenMessage<UploadPartRequest>;

/**
 * @generated from message acyclic.objects.v1.UploadedPart
 */
export declare type UploadedPart = Message<"acyclic.objects.v1.UploadedPart"> & {
  /**
   * @generated from field: uint32 part_number = 1;
   */
  partNumber: number;

  /**
   * @generated from field: string etag = 2;
   */
  etag: string;

  /**
   * @generated from field: uint64 size = 3;
   */
  size: bigint;
};

/**
 * Describes the message acyclic.objects.v1.UploadedPart.
 * Use `create(UploadedPartSchema)` to create a new message.
 */
export declare const UploadedPartSchema: GenMessage<UploadedPart>;
// ...
export declare const ListPartsRequestSchema: GenMessage<ListPartsRequest>;
// ...
export declare const ListPartsResponseSchema: GenMessage<ListPartsResponse>;
// ...
export declare const CompleteMultipartRequestSchema: GenMessage<CompleteMultipartRequest>;
// ...
export declare const AbortMultipartRequestSchema: GenMessage<AbortMultipartRequest>;
// ...
export declare const AbortMultipartResponseSchema: GenMessage<AbortMultipartResponse>;
// ...
export declare type ErrorDetail = Message<"acyclic.objects.v1.ErrorDetail"> & {
  /**
   * @generated from field: acyclic.objects.v1.ErrorCode code = 1;
   */
  code: ErrorCode;

  /**
   * @generated from field: string request_id = 2;
   */
  requestId: string;
};

/**
 * Describes the message acyclic.objects.v1.ErrorDetail.
 * Use `create(ErrorDetailSchema)` to create a new message.
 */
export declare const ErrorDetailSchema: GenMessage<ErrorDetail>;
// ...
export enum ErrorCode {
  /**
   * @generated from enum value: ERROR_CODE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: ERROR_CODE_INVALID_ARGUMENT = 1;
   */
  INVALID_ARGUMENT = 1,

  /**
   * @generated from enum value: ERROR_CODE_NOT_FOUND = 2;
   */
  NOT_FOUND = 2,

  /**
   * @generated from enum value: ERROR_CODE_ALREADY_EXISTS = 3;
   */
  ALREADY_EXISTS = 3,

  /**
   * @generated from enum value: ERROR_CODE_PRECONDITION_FAILED = 4;
   */
  PRECONDITION_FAILED = 4,

  /**
   * @generated from enum value: ERROR_CODE_IDEMPOTENCY_MISMATCH = 5;
   */
  IDEMPOTENCY_MISMATCH = 5,

  /**
   * @generated from enum value: ERROR_CODE_TOKEN_EXPIRED = 6;
   */
  TOKEN_EXPIRED = 6,

  /**
   * @generated from enum value: ERROR_CODE_QUOTA_EXCEEDED = 7;
   */
  QUOTA_EXCEEDED = 7,

  /**
   * @generated from enum value: ERROR_CODE_UNSUPPORTED = 8;
   */
  UNSUPPORTED = 8,

  /**
   * @generated from enum value: ERROR_CODE_UNAVAILABLE = 9;
   */
  UNAVAILABLE = 9,
}

/**
 * Describes the enum acyclic.objects.v1.ErrorCode.
 */
export declare const ErrorCodeSchema: GenEnum<ErrorCode>;

/**
 * @generated from service acyclic.objects.v1.BucketsService
 */
export declare const BucketsService: GenService<{
  /**
   * @generated from rpc acyclic.objects.v1.BucketsService.CreateBucket
   */
  createBucket: {
    methodKind: "unary";
    input: typeof CreateBucketRequestSchema;
    output: typeof BucketSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.BucketsService.HeadBucket
   */
  headBucket: {
    methodKind: "unary";
    input: typeof HeadBucketRequestSchema;
    output: typeof BucketSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.BucketsService.DeleteBucket
   */
  deleteBucket: {
    methodKind: "unary";
    input: typeof DeleteBucketRequestSchema;
    output: typeof DeleteBucketResponseSchema;
  },
}>;

/**
 * @generated from service acyclic.objects.v1.ObjectsService
 */
export declare const ObjectsService: GenService<{
  /**
   * @generated from rpc acyclic.objects.v1.ObjectsService.PutObject
   */
  putObject: {
    methodKind: "client_streaming";
    input: typeof PutObjectRequestSchema;
    output: typeof ObjectVersionSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.ObjectsService.GetObject
   */
  getObject: {
    methodKind: "server_streaming";
    input: typeof GetObjectRequestSchema;
    output: typeof GetObjectResponseSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.ObjectsService.HeadObject
   */
  headObject: {
    methodKind: "unary";
    input: typeof HeadObjectRequestSchema;
    output: typeof HeadObjectResponseSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.ObjectsService.DeleteObject
   */
  deleteObject: {
    methodKind: "unary";
    input: typeof DeleteObjectRequestSchema;
    output: typeof DeleteObjectResponseSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.ObjectsService.ListObjects
   */
  listObjects: {
    methodKind: "unary";
    input: typeof ListObjectsRequestSchema;
    output: typeof ListObjectsResponseSchema;
  },
}>;

/**
 * @generated from service acyclic.objects.v1.MultipartService
 */
export declare const MultipartService: GenService<{
  /**
   * @generated from rpc acyclic.objects.v1.MultipartService.CreateMultipart
   */
  createMultipart: {
    methodKind: "unary";
    input: typeof CreateMultipartRequestSchema;
    output: typeof MultipartUploadSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.MultipartService.UploadPart
   */
  uploadPart: {
    methodKind: "client_streaming";
    input: typeof UploadPartRequestSchema;
    output: typeof UploadedPartSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.MultipartService.ListParts
   */
  listParts: {
    methodKind: "unary";
    input: typeof ListPartsRequestSchema;
    output: typeof ListPartsResponseSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.MultipartService.CompleteMultipart
   */
  completeMultipart: {
    methodKind: "unary";
    input: typeof CompleteMultipartRequestSchema;
    output: typeof ObjectVersionSchema;
  },
  /**
   * @generated from rpc acyclic.objects.v1.MultipartService.AbortMultipart
   */
  abortMultipart: {
    methodKind: "unary";
    input: typeof AbortMultipartRequestSchema;
    output: typeof AbortMultipartResponseSchema;
  },
}>;